chore[cli] Moved common cdk actions to utils/base class
